refactor: update entity package structure and remove auto-generate from primary keys
