+++ local_datasource/src/main/java/com/baghdad/local_datasource/database/dto/RecentSearch.kt
@@ -1,12 +1,12 @@
-package com.baghdad.local_datasource.roomDB.entity
+package com.baghdad.local_datasource.database.dto
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.baghdad.local_datasource.roomDB.entity.RecentSearch.Companion.RECENTLY_VIEWED_TABLE_NAME
+import com.baghdad.local_datasource.database.dto.RecentSearch.Companion.RECENTLY_SEARCH_TABLE_NAME
 import com.baghdad.repository.model.RecentSearchDto
 
-@Entity(tableName = RECENTLY_VIEWED_TABLE_NAME)
+@Entity(tableName = RECENTLY_SEARCH_TABLE_NAME)
 data class RecentSearch(
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = ID)
@@ -17,7 +17,7 @@ data class RecentSearch(
     val searchedAt: Long = System.currentTimeMillis()
 ) {
     companion object {
-        const val RECENTLY_VIEWED_TABLE_NAME = "RecentlyViewed"
+        const val RECENTLY_SEARCH_TABLE_NAME = "RecentlySearch"
         const val ID = "id"
         const val QUERY = "query"
         const val SEARCHED_AT = "searchedAt"
